--- prompt/test/integration/input_no_location_flag_test.ts
@@ -1,17 +1,0 @@
-import { Input } from "../../input.ts";
-import { assertSnapshotCall } from "../../../testing/assert_snapshot_call.ts";
-
-await assertSnapshotCall({
-  name: "input prompt with no location flag",
-  meta: import.meta,
-  osSuffix: ["windows"],
-  steps: {
-    "should work without --location flag": { stdin: ["yes", "\n"] },
-  },
-  async fn() {
-    await Input.prompt({
-      message: "Works without --location?",
-      default: "hope so",
-    });
-  },
-});
