--- prompt/test/integration/input_no_prefix_test.ts
@@ -1,18 +1,0 @@
-import { Input } from "../../input.ts";
-import { assertSnapshotCall } from "../../../testing/assert_snapshot_call.ts";
-
-await assertSnapshotCall({
-  name: "input prompt without prefix",
-  meta: import.meta,
-  osSuffix: ["windows"],
-  steps: {
-    "should disable prefix": { stdin: ["bar", "\n"] },
-  },
-  async fn() {
-    await Input.prompt({
-      message: "Whats your name?",
-      default: "foo",
-      prefix: "",
-    });
-  },
-});
