+++ prompt/test/integration/number_test.ts
@@ -32,3 +32,30 @@ await assertSnapshotCall({
     });
   },
 });
+
+await assertSnapshotCall({
+  name: "number prompt with float",
+  meta: import.meta,
+  osSuffix: ["windows"],
+  steps: {
+    "should enter a floating number": {
+      stdin: ansi
+        .text("19.")
+        .cursorUp
+        .cursorUp
+        .cursorUp
+        .cursorUp
+        .text("\n")
+        .toArray(),
+    },
+  },
+  async fn() {
+    await Number.prompt({
+      message: "How old are you?",
+      default: 7,
+      max: 20,
+      min: 18,
+      float: true,
+    });
+  },
+});
